# backend/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/instagram/scheduling/test-publish/{postId}": {
      "post": {
        "operationId": "InstagramSchedulingController_testPublish",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Scheduled post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test publish result with detailed logs"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Test publish a scheduled post (DEBUG)",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/posts": {
      "post": {
        "operationId": "InstagramSchedulingController_createScheduledPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post scheduled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledPostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create scheduled post",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/posts/{accountId}": {
      "get": {
        "operationId": "InstagramSchedulingController_listScheduledPosts",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Client account ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "scheduled",
                "publishing",
                "published",
                "failed",
                "cancelled"
              ]
            }
          },
          {
            "name": "scheduledAfter",
            "required": false,
            "in": "query",
            "description": "Get posts scheduled after this date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scheduledBefore",
            "required": false,
            "in": "query",
            "description": "Get posts scheduled before this date",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedScheduledPostsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List scheduled posts",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/posts/{accountId}/{postId}": {
      "get": {
        "operationId": "InstagramSchedulingController_getScheduledPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Scheduled post ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Client account ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Post retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledPostResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get scheduled post by ID",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/posts/{postId}": {
      "put": {
        "operationId": "InstagramSchedulingController_updateScheduledPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Scheduled post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduledPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledPostResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot update non-scheduled posts"
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update scheduled post",
        "tags": [
          "Instagram Scheduling"
        ]
      },
      "delete": {
        "operationId": "InstagramSchedulingController_cancelScheduledPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Scheduled post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post cancelled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelScheduledPostResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel scheduled post",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/posts/{postId}/publish-now": {
      "post": {
        "operationId": "InstagramSchedulingController_publishNow",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Scheduled post ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post published successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishNowResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Post not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish scheduled post immediately",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/publish": {
      "post": {
        "operationId": "InstagramSchedulingController_publishInstantly",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduledPostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post published successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishNowResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Publish post immediately without scheduling",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/media/upload": {
      "post": {
        "operationId": "InstagramSchedulingController_uploadMedia",
        "parameters": [
          {
            "name": "clientAccountId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Media uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid file"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload media file",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/media/{accountId}": {
      "get": {
        "operationId": "InstagramSchedulingController_listMedia",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Client account ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "mediaType",
            "required": false,
            "in": "query",
            "description": "Filter by media type",
            "schema": {
              "type": "string",
              "enum": [
                "image",
                "video"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media assets retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaAssetsListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List media assets",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/media/{accountId}/{assetId}": {
      "get": {
        "operationId": "InstagramSchedulingController_getMediaAsset",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "description": "Media asset ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Client account ID",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Media asset retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Media asset not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get media asset details",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/media/{assetId}": {
      "delete": {
        "operationId": "InstagramSchedulingController_deleteMediaAsset",
        "parameters": [
          {
            "name": "assetId",
            "required": true,
            "in": "path",
            "description": "Media asset ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media asset deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMediaAssetResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete media in use"
          },
          "404": {
            "description": "Media asset not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete media asset",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
    "/api/instagram/scheduling/optimal-times/{accountId}": {
      "get": {
        "operationId": "InstagramSchedulingController_getOptimalTimes",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "description": "Client account ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Optimal times retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimalPostingTimesResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Account not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get optimal posting times",
        "tags": [
          "Instagram Scheduling"
        ]
      }
    },
...
  },
  "info": {
    "title": "Social Selling API",
    "description": "\n# Social Selling Platform API\n\nComplete REST API for Instagram-based social selling platform.\n\n## Features\n- Instagram OAuth integration\n- Product catalog management\n- Direct message automation\n- Post scheduling and analytics\n- Multi-account support\n\n## Authentication\nAll authenticated endpoints require a Bearer token in the Authorization header.\nObtain tokens via POST /api/auth/login endpoint.\n\n## Rate Limiting\n- Authenticated: 1000 requests/hour\n- Unauthenticated: 100 requests/hour\n\n## Session Management\nSessions are managed via HTTP-only cookies and can be tracked per device.\n  ",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local Development"
    },
    {
      "url": "https://staging-api.socialselling.willianbvsanches.com",
      "description": "Staging Environment"
    },
    {
      "url": "https://api.socialselling.willianbvsanches.com",
      "description": "Production Environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT access token",
        "in": "header"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "ssell_session",
        "description": "Session cookie for authenticated requests"
      }
    },
    "schemas": {
...
      "ProductTagDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "Product ID"
          },
          "x": {
            "type": "number",
            "description": "X coordinate (0-1)"
          },
          "y": {
            "type": "number",
            "description": "Y coordinate (0-1)"
          }
        },
        "required": [
          "productId",
          "x",
          "y"
        ]
      },
      "CreateScheduledPostDto": {
        "type": "object",
        "properties": {
          "clientAccountId": {
            "type": "string",
            "description": "Client account ID"
          },
          "scheduledFor": {
            "type": "string",
            "description": "Scheduled publish time (ISO 8601)"
          },
          "caption": {
            "type": "string",
            "description": "Post caption"
          },
          "mediaUrls": {
            "description": "Media URLs (uploaded to S3)",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mediaType": {
            "type": "string",
            "description": "Media type",
            "enum": [
              "IMAGE",
              "VIDEO",
              "CAROUSEL",
              "REELS",
              "STORIES"
            ]
          },
          "productTags": {
            "description": "Product tags for shopping posts",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductTagDto"
            }
          },
          "locationId": {
            "type": "string",
            "description": "Location ID"
          },
          "templateId": {
            "type": "string",
            "description": "Template ID to use"
          },
          "templateVariables": {
            "type": "object",
            "description": "Template variables"
          }
        },
        "required": [
          "clientAccountId",
          "scheduledFor",
          "caption",
          "mediaUrls",
          "mediaType"
        ]
      },
      "ScheduledPostResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientAccountId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "scheduledFor": {
            "format": "date-time",
            "type": "string"
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string"
          },
          "caption": {
            "type": "string"
          },
          "mediaUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "IMAGE",
              "VIDEO",
              "CAROUSEL",
              "REELS",
              "STORIES"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "publishing",
              "published",
              "failed",
              "cancelled"
            ]
          },
          "publishAttempts": {
            "type": "number"
          },
          "lastPublishError": {
            "type": "string"
          },
          "instagramMediaId": {
            "type": "string"
          },
          "permalink": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "cancelledAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "clientAccountId",
          "userId",
          "scheduledFor",
          "caption",
          "mediaUrls",
          "mediaType",
          "status",
          "publishAttempts",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedScheduledPostsResponseDto": {
        "type": "object",
        "properties": {
          "posts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScheduledPostResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "posts",
          "total",
          "page",
          "limit"
        ]
      },
      "UpdateScheduledPostDto": {
        "type": "object",
        "properties": {
          "scheduledFor": {
            "type": "string",
            "description": "Scheduled publish time"
          },
          "caption": {
            "type": "string",
            "description": "Post caption"
          },
          "mediaUrls": {
            "description": "Media URLs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "productTags": {
            "description": "Product tags",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductTagDto"
            }
          },
          "locationId": {
            "type": "string",
            "description": "Location ID"
          }
        }
      },
      "CancelScheduledPostResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "scheduledPostId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "scheduledPostId",
          "status"
        ]
      },
      "PublishNowResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "scheduledPostId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "scheduledPostId",
          "status"
        ]
      },
      "MediaUploadResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "originalFilename": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "fileSize": {
            "type": "number"
          },
          "mediaType": {
            "type": "string"
          },
          "s3Url": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          },
          "width": {
            "type": "number"
          },
          "height": {
            "type": "number"
          },
          "duration": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "filename",
          "originalFilename",
          "mimeType",
          "fileSize",
          "mediaType",
          "s3Url",
          "createdAt"
        ]
      },
      "MediaAssetsListResponseDto": {
        "type": "object",
        "properties": {
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaUploadResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "assets",
          "total",
          "page",
          "limit"
        ]
      },
      "DeleteMediaAssetResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "mediaId": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "mediaId"
        ]
      },
...
    }
  }
}